[WFLY-5678] Do not migrate attributes that no longer accept expressions
The discovery-group and broadcast-group resources in the legacy
messaging subsystem accept expressions for their jgroups-stack and
jgroups-channel resources.
The same resources in the new messaging-activemq subsystem no longer
accepts expressions for them (as they are references to other model
resources).

During migration, these attributes are ignored and removed from the new
:add operations to create the messaging-activemq resources and warnings
are added to the migration-warnings.
This allows the :migrate operation to succeed and the user will have to
fix the model afterwards by specifying a correct value for these
attributes.

JIRA: https://issues.jboss.org/browse/WFLY-5678