[WFLY-5621] Cleanup messaging XSD schemas
Remove the jboss-as-messaging-deployment_1_0.xsd schema that is no
longer used. The messaging subsystem has no runtime and these files can
not be used anymore by deployments.

The others XSD files are kept as it is possible to use them to edit
messaging subsystem XML configurations (either for legacy slaves or for
admin-only server prior to migration).

JIRA: https://issues.jboss.org/browse/WFLY-5621