[JBEAP-2573] jboss-as-messaging_1_4.xsd in EAP7 is not the same as jboss-as-messaging_1_4.xsd in EAP6
Copy jboss-as-messaging_1_4.xsd from EAP6 to EAP7 to ensure that they
are the same schema (1.4 is the latest schema version in EAP6).

The 1.4 XSD differs in WildFly (as explained in WFLY-4533). With this
commit we end up with having the 1.4 XSD in EAP6 & 7 correspond to the
1.5 XSD in WFLY 8/9/10.
Note that the 1.4 parsers was updated to be able to parse both 1.4 and
1.5 XML documents.

JIRA: https://issues.jboss.org/browse/JBEAP-2573