Use the defined authentication-context and recovery-authentication-context for the ElytronSubjectFactory on data sources
Add the authentication-context dependency to the ConnectionDefinitionService. For this simple approach the dependency is added to guard against the authentication-context being removed if it's in use by a resource-adapter. This should later be moved into a proper capability itself so the value of the authentication-context capability can be directly used.