[WFLY-8246] Static cluster with http-connector
add server-name attribute to the http-connector resources so that the
connector can connect to the correct ActiveMQ Artemis server on a remote
app server (as the app server can embed multiple Artemis server with the
same acceptor names, the Artemis server name is required to identifed
the correct one).
If this attribute is undefined, the http-connector will use the name of
its parent Artemis server (making it suitable for http-connector used to
connect to the server itself).

JIRA: https://issues.jboss.org/browse/WFLY-8246