[JBEAP-8291] fix for TxTestUtil works with security manager
Main point was not using CDI.current() as for such thing security
manager would need very special settings, see
https://github.com/wildfly/wildfly/pull/9591https://github.com/wildfly/wildfly/pull/9591

There is an additional fix for using adjust for MDB timeout tests as
was changed by WFLY-8109.
There was an issue in test which does not wait long enough to
receive a message.