[WFLY-8593] Wildfly should log bound ports
When specifying that Wildfly should bind to an ephemeral port (by
specifying the port number as 0), Wildfly dutifully logs that it is
bound to port 0. Yet, this is not right as instead the networking layer
determined an actual port number when binding the socket. Instead,
Wildfly should log the actual port that it is bound to. This commit
causes this to be the case.