[WFLY-8746] RejectedExecutionException in the MDB during shutdown
The PooledConnectionFactoryService was is used by MDB was not depending
on the service controlling the lifecycle of Artemis client thread pools.
It was possible that the thread pools were cleaned (and stopped) before
the MDB was stopped causing RejectedExecutionException during MDB code
execution.

JIRA: https://issues.jboss.org/browse/WFLY-8746