JBEAP-8937 - build.sh and integration-tests.sh scripts doesn't work on Solaris 10 (SPARC)
Turns out previous fix did not solve the core issue - the bash interpreter used for testing has a bug, that cannot be workaround in a predictable fashion. To workaround the issue, the best approach is to provided a way to use an other interpreter. As the build.sh scripts has '/bin/sh' hardcoded in it, this commit provides a mechanism to override the path to the interpreter if needed.