[WFLY-8928] Allow local transactions for JMS RA session
By default, a JMS session created from a RA does not allow local
transactions (as specified by the JMS 2.0 API).
However, for legacy compatibility, this behaviour can be changed by
setting the allow-local-transactions attribute to true on the
pooled-connection-factory resource.

During migration of the legacy messaging subsystem, this attribute
is set to true to preserve the behaviour of HornetQ RA that was
allowing local transactions.

JIRA: https://issues.jboss.org/browse/WFLY-8928